refactor : domain 패키지 없어진거 다시 생성
