+++ src/test/java/laddergame/domain/LadderTest.java
@@ -1,8 +1,5 @@
 package laddergame.domain;
 
-import laddergame.Height;
-import laddergame.Ladder;
-import laddergame.Players;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
