+++ src/test/java/laddergame/domain/PlayersTest.java
@@ -1,6 +1,5 @@
 package laddergame.domain;
 
-import laddergame.Players;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
